scripts/gcal/requirements.txt
python-dateutil
google-api-python-client
oauth2client
icalendar